refactor(analysis-types): split corporate DTOs
