+++ crates/tokmd-analysis-types/src/corporate.rs
@@ -1,0 +1,13 @@
+use serde::{Deserialize, Serialize};
+
+#[derive(Debug, Clone, Serialize, Deserialize)]
+pub struct CorporateFingerprint {
+    pub domains: Vec<DomainStat>,
+}
+
+#[derive(Debug, Clone, Serialize, Deserialize)]
+pub struct DomainStat {
+    pub domain: String,
+    pub commits: u32,
+    pub pct: f32,
+}
