+++ crates/tokmd-analysis-types/src/git.rs
@@ -2,22 +2,6 @@ use serde::{Deserialize, Serialize};
 
 use crate::churn::TrendClass;
 
-// ---------------------
-// Corporate fingerprint
-// ---------------------
-
-#[derive(Debug, Clone, Serialize, Deserialize)]
-pub struct CorporateFingerprint {
-    pub domains: Vec<DomainStat>,
-}
-
-#[derive(Debug, Clone, Serialize, Deserialize)]
-pub struct DomainStat {
-    pub domain: String,
-    pub commits: u32,
-    pub pct: f32,
-}
-
 // ---------
 // Git report
 // ---------
